requirements.txt
# file: requirements.txt

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Babel==4.0.0

Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0

python-dotenv==1.0.1

requests==2.31.0

stripe==10.12.0

gunicorn==21.2.0

aiogram==3.13.1
